py/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.15.0
eth-abi>=4.2.1
eth-utils>=2.3.1
PyNaCl>=1.5.0
base58>=2.1.1
flask>=3.0.0
flask-cors>=4.0.0

